chore: move arquivos da pasta ui para lib
