+++ lib/widgets/news_item_card.dart
