+++ lib/widgets/shop_card.dart
