+++ lib/widgets/tabs/news_card_tab.dart
